test: convert cypress test setup to 100% typescript
