+++ tests/cypress/.eslintrc
@@ -4,6 +4,7 @@
   ],
   "rules": {
     "@typescript-eslint/no-var-requires": "off",
+    "@typescript-eslint/no-namespace": "off",
     "import/no-unresolved": [
       "error",
       {
