+++ tests/cypress/component/hoverOverlay.spec.tsx
