+++ tests/cypress/component/overlayTransitionDuration.spec.tsx
