+++ tests/cypress/component/pausedOverlay.spec.tsx
