+++ tests/cypress/component/playbackRangeEnd.spec.tsx
