+++ tests/cypress/component/playbackRangeStartAndEnd.spec.tsx
