+++ tests/cypress/component/restartOnPaused.spec.tsx
