+++ tests/cypress/component/styleProps.spec.tsx
