+++ tests/cypress/cypress.json
@@ -2,10 +2,10 @@
   "componentFolder": "tests/cypress/component",
   "fixturesFolder": "tests/cypress/fixtures",
   "integrationFolder": "tests/cypress/integration",
-  "pluginsFile": "tests/cypress/plugins/index.js",
+  "pluginsFile": "tests/cypress/plugins/index.ts",
   "screenshotsFolder": "tests/cypress/screenshots",
   "videosFolder": "tests/cypress/videos",
-  "supportFile": "tests/cypress/support/index.js",
+  "supportFile": "tests/cypress/support/index.ts",
   "testFiles": "**/*.spec.{js,ts,jsx,tsx}",
   "video": false,
   "screenshotOnRunFailure": false,
